requirements.txt
flask
transformers
sentence-transformers
elasticsearch
openai-whisper
requests
